Move withCachedData to components package
