+++ packages/components/src/hooks/withCachedData.ts
@@ -1,6 +1,8 @@
-import { AutoLoadedState, useLoadedDataEvents } from '@terra-ui-packages/components';
 import { Atom, LoadedState, ReadyState } from '@terra-ui-packages/core-utils';
-import { useSettableStore } from 'src/libs/react-utils';
+
+import { AutoLoadedState } from './useAutoLoadedData';
+import { useLoadedDataEvents } from './useLoadedDataEvents';
+import { useSettableStore } from './useStore';
 
 type LoadedDataFn<S, E = unknown> = (
   ...args: any[]
