+++ packages/components/src/index.ts
@@ -16,8 +16,10 @@ export * from './hooks/useAutoLoadedData';
 export * from './hooks/useLoadedData';
 export * from './hooks/useLoadedDataEvents';
 export * from './hooks/usePrevious';
+export * from './hooks/useStore';
 export * from './hooks/useUniqueId';
 export * from './hooks/useWindowDimensions';
+export * from './hooks/withCachedData';
 export * from './icon';
 export type { IconId } from './icon-library';
 export * from './injectStyle';
