+++ src/libs/react-utils.test.ts
@@ -1,7 +1,6 @@
-import { Atom, atom } from '@terra-ui-packages/core-utils';
 import { act, renderHook } from '@testing-library/react';
 
-import { useDebouncedValue, useSettableStore } from './react-utils';
+import { useDebouncedValue } from './react-utils';
 
 describe('useDebouncedValue', () => {
   it('debounces changes to value', () => {
@@ -32,50 +31,5 @@ describe('useDebouncedValue', () => {
     expect(valueAfterUpdate).toBe('foo');
     expect(valueAfte500ms).toBe('foo');
     expect(valueAfter1000ms).toBe('bar');
-  });
-});
-
-describe('useSettableStore', () => {
-  it('handles direct value update', () => {
-    // Arrange
-    const myAtom = atom<string>('hello');
-    const { result: hookReturnRef, rerender } = renderHook((args: [Atom<string>]) => useSettableStore(...args), {
-      initialProps: [myAtom],
-    });
-    const [initialValue, setValue] = hookReturnRef.current;
-
-    // Act
-    act(() => {
-      setValue('goodbye');
-    });
-    rerender([myAtom]);
-
-    // Assert
-    expect(initialValue).toBe('hello');
-    const [value] = hookReturnRef.current;
-    expect(value).toBe('goodbye');
-    expect(myAtom.get()).toBe('goodbye');
-  });
-
-  it('reacts to updated atom value', () => {
-    // Arrange
-    const myAtom = atom<string>('hello');
-    const { result: hookReturnRef, rerender } = renderHook((args: [Atom<string>]) => useSettableStore(...args), {
-      initialProps: [myAtom],
-    });
-
-    const [initialValue] = hookReturnRef.current;
-
-    // Act
-    act(() => {
-      myAtom.set('goodbye');
-    });
-    rerender([myAtom]);
-
-    // Assert
-    expect(initialValue).toBe('hello');
-    const [value] = hookReturnRef.current;
-    expect(value).toBe('goodbye');
-    expect(myAtom.get()).toBe('goodbye');
   });
 });
