+++ src/libs/react-utils.ts
@@ -1,4 +1,4 @@
-import { Atom, delay, safeCurry } from '@terra-ui-packages/core-utils';
+import { delay, safeCurry } from '@terra-ui-packages/core-utils';
 import _ from 'lodash/fp';
 import {
   EffectCallback,
@@ -12,6 +12,8 @@ import {
 } from 'react';
 import { h } from 'react-hyperscript-helpers';
 import { pollWithCancellation } from 'src/libs/utils';
+
+export { useStore } from '@terra-ui-packages/components';
 
 /**
  * Performs the given effect, but only on component mount.
@@ -169,29 +171,6 @@ export const useCurrentTime = (initialDelay = 250) => {
       delayRef.current = delay;
     },
   ];
-};
-
-/**
- * Hook that returns the value of a given store. When the store changes, the component will re-render
- */
-export const useStore = <T>(theStore: Atom<T>): T => {
-  const [value, setValue] = useState(theStore.get());
-  useEffect(() => {
-    return theStore.subscribe((v) => setValue(v)).unsubscribe;
-  }, [theStore]);
-  return value;
-};
-
-/**
- * Hook that returns a tuple with [currentValue, setValueFn] for a given store.  This provides an alternative
- * to useStore hook above that more closely emulates the tuple return of built-in useState hook.
- * When the store changes, the component will re-render.
- * @param theStore
- */
-export const useSettableStore = <T>(theStore: Atom<T>): [T, (value: T) => void] => {
-  const currentStoreValue = useStore(theStore);
-
-  return [currentStoreValue, theStore.set];
 };
 
 export const useDebouncedValue = <T>(value: T, wait: number): T => {
